functions/transcribe-py/requirements.txt
appwrite
deepgram-sdk
